First attempt at gettting chromedriver2 to work on android.
This is for the configuration where chromedriver is running on the host.
From chromedriver, we invoke adb commands to clear the profile, launch
the app, and forward the host port to the named abstract socket on the
device. For now, we assume a single device is connected.

Also the following refactoring:
- Fold ChromeLauncher into Chrome
- Create a base ChromeImpl which desktop and android implementations
  are derived from

BUG=166722


Review URL: https://chromiumcodereview.appspot.com/12082129

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181041 0039d316-1c4b-4281-b951-d872f2087c98