Reland r212020: Move Python setup/tear down logic into Forwarder ...
Forwarder used to be a pain to setup/tear down across all the various
harnesses.
This CL should hopefully solve these issues by hiding these
implementation details. The host daemon is now killed once the first time that
the Forwarder class is used and the daemon running on the devices is also
killed the first time a port is forwarded for a specific device.

TBR=bulach@chromium.org
BUG=242846

Review URL: https://codereview.chromium.org/19550004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@212582 0039d316-1c4b-4281-b951-d872f2087c98