[chromedriver] Use document.readyState to determine loading state for M44+.
As of crrev.com/323900 (Chrome 44+) we no longer receive
Page.frameStartedLoading and Page.frameStoppedLoading events immediately after
clicking on an element. This can cause tests to fail if the click command
results in a page navigation, such as in the case of
PageLoadingTest.testShouldTimeoutIfAPageTakesTooLongToLoadAfterClick from the
Selenium test suite.

Unfortunately, this breaks
PageLoadingTest.testShouldNotHangIfDocumentOpenCallIsNeverFollowedByDocumentCloseCall,
which I've also disabled in this CL. A long-term fix will require a new DevTools
command to query the page loading state.

BUG=chromedriver:1158,chromedriver:1167

Review URL: https://codereview.chromium.org/1250333002

Cr-Commit-Position: refs/heads/master@{#340529}