[chromedriver] When tracking execution contexts, ignore dummy frames.
This reduces flakiness for tests which create a new window and switch to it
mid-way through the page load. Recently, testCanLaunchApp from run_py_tests.py
has been flaky for this reason.

The problem occurs when we switch to a new window that is part-way through the
page load. When the page loading state is unknown, we append a dummy frame to
the page and observe the resulting DevTools events. When the dummy frame is
deleted, so is it's execution context, and if this is the only execution context
that ChromeDriver knows about, it will think a new navigation is pending and
hang.

BUG=chromedriver:1167
TBR=stgao@chromium.org

Review URL: https://codereview.chromium.org/1287823002

Cr-Commit-Position: refs/heads/master@{#342971}