[chromedriver] Apply page load timeout to slow cross-process navigations
This is a second attempt to fight chromedriver hangs on slow-to-commit
cross-process navigations. In the discussion of the first, incorrect
solution (see https://codereview.chromium.org/1520883002/), it became
clear that it'd be better to handle this in chromedriver. Here I
introduce a timeout tracking mechanism which is easy to use in the
chromedriver's architecture. In general, it allows to abort waiting for
a devtools command response, do some cleanup and return an error to the
client.

Currently this timeout tracking is used only when executing the 'get'
command, history navigations and reloads are left as is. I wanted to get
feedback about the approach as a whole first. Please take a look.

BUG=chromedriver:1292

Review URL: https://codereview.chromium.org/1669453002

Cr-Commit-Position: refs/heads/master@{#387658}