Change error pages to use a new chrome-error:// scheme.
This CL changes error pages to use chrome-error://chromewebdata/
instead of data:text/html,chromewebdata.

This has the benefit that error pages won't inherit CSP from their
parent/opener.  Previously, when a page defined a CSP and then ended
up loading an error page in a subframe, the error page might've been
broken by the parent's CSP, since, for example, it needs to execute
inline scripts.  This could result in the error page not showing up
correctly and/or false CSP reports being sent.

The new scheme is marked as secure and as requiring an opaque origin
to match previous behavior.

Web pages used to be able to directly load the error URL, which just
showed up as "chromewebdata".  With this change, navigating to the
error URL would bring up the external protocol dialog instead, so this
CL prevents renderers from directly navigating to or redirecting to
error URLs.

Additionally, chrome-error:// is registered as a display-isolated
scheme, so that regular web pages can't embed the error URL in an
iframe or image, and as a scheme that does not allow javascript URL
manipulation, which is consistent with other pages considered to be
part of Chrome.  If either of these new restrictions ends up being
problematic, we should revisit them in
RenderThreadImpl::RegisterSchemes().

In the future, it's possible to further utilize the host/path portion
of the URL to identify different kinds of error pages.

Bug: 703801
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_site_isolation
Change-Id: I45fa0aa157523450c33c6464f96414ff742e8604
Reviewed-on: https://chromium-review.googlesource.com/580169
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Charlie Reis (OOO Aug 17-24) <creis@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#494809}