[ChromeDriver] Fix version reported by ChromeDriver
ChromeDriver reports the revision number and Git hash of the source code
it is built from, but the reported value is usually a little off. This
is fixed by making ChromeDriver's version.cc depend on LASTCHANGE,
which is updated by gclient whenever the source tree is sync'ed to a
different Git revision.

Change-Id: I811ba9912f5c74a04e0fa18b95dfc5fd61f7bd28
Reviewed-on: https://chromium-review.googlesource.com/861153
Reviewed-by: Jonathon Kereliuk <kereliuk@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#528659}