Reland "[ChromeDriver] Enable CQ tests on Mac/Windows"
This is a reland of f3cf593451c3bf68d3ea643029a84fa2537dbf08,
with the change to enable the ChromeDriver tests on release bot only.
The original commit was reverted due to flakiness of the tests on debug
build. The reason for the flakiness is debug build of Chrome takes a
long time to start, usually between 20 to 40 seconds. The timeout for
the test is 30 seconds, thus it can either succeed or fail depending
on whether Chrome started quick enough or not.

While it is possible to increase the timeout, spending over 30 seconds
per test appears to be too much overhead in the long run. Also, the
purpose of running these tests is to catch incompatibilities that can
develop between Chrome and ChromeDriver, and running these tests on
release builds should suffice for that purpose.

Original change's description:
> [ChromeDriver] Enable CQ tests on Mac/Windows
>
> Previously, the commit queue only runs ChromeDriver tests on Linux.
> This CL expands the test to other desktop platforms, including Mac
> and Windows.
>
> Bug: 793370
> Change-Id: I2a28d3a5c8371edc28566010739681acfac41c21
> Reviewed-on: https://chromium-review.googlesource.com/881641
> Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
> Reviewed-by: John Budorick <jbudorick@chromium.org>
> Commit-Queue: John Chen <johnchen@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#531399}

Bug: 793370
Change-Id: I87df556896593ab9497435611a7d78122adbddd8
Reviewed-on: https://chromium-review.googlesource.com/890391
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#532924}