[ChromeDriver] Include test name in Python test log
The ChromeDriver Python test script run_py_tests.py runs a large number
of tests, and when a test fails, it is hard to find the right section in
the log file. This CL adds the test name in the log file to aid
debugging. It uses a non-standard capability goog:testName, which is
included in the log file but otherwise ignored by ChromeDriver.

Change-Id: Ia63e9f11d903ddd363442fce903f54d8c1070a07
Reviewed-on: https://chromium-review.googlesource.com/902173
Reviewed-by: Jonathon Kereliuk <kereliuk@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#534505}