[ChromeDriver] Allow user-defined remote-debugging-port.
Chromium Embedded Framework sometimes uses an in-memory user-data-dir,
so the DevToolsActivePort file is not a viable option for them. We
think it is likely that other Chromium variants, present or future,
may have trouble with the DevToolsActivePort file, so this should
make things easier for them.

Also, fix testConnectToRemoteBrowser to
1. Have retries so it passes in the case of a race condition
for port allocation.
2. Prevent a race with Chrome during teardown.
   -- Note that this race was causing the user-data-dir to
      not be deleted at the end.

Bug: 863247
Change-Id: I506be26fc2e282b7194afd88afd02efcec65e880
Reviewed-on: https://chromium-review.googlesource.com/1137440
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#575408}