[ChromeDriver] Fix test error caused by W3C flag change
Part of testW3cCompliantResponses is failing on ChromeDriver waterfall,
due to an unusual setting (W3C flag set to true in a legacy format Init
Session command). This test works if there is a default Chrome
installation (e.g., dev machine) or if Chrome is located in the same
directory as ChromeDriver (e.g., Commit Queue), but fails if Chrome is
at a different location (e.g., ChromeDriver waterfall). Removing the
problematic setting as it's unexpected in real world.

TBR=crouleau@chromium.org

Change-Id: I086c772c6c7521f643739be5ff4485c05fad760a
Reviewed-on: https://chromium-review.googlesource.com/1227714
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#591592}