[ChromeDriver] Modify ExecutePerformActions command
In order to match with WebDriver Action API, we need to modify the
command of ExecutePerformActions in ChromeDriver. I also add this new
action API in chromederiver.py and some tests to test this new API.

Because we have not added this Action API to CDP, I am using
Input.dispatchMouseEvent right now for mouse events only.

Bug: chromedriver:1897
Change-Id: I28cac07b58d570ab1a102ac7b13d3d06dac507d8
Reviewed-on: https://chromium-review.googlesource.com/c/1333831
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#611732}