Drop isomorphic decoding from fragment navigation
When Chrome opens a URL with a fragment id, it decodes %xx for UTF-8,
and decodes it for isomorphic-decode [1] if it is invalid for UTF-8.
Then tries to find an element with the decoding result as an ID.

No other browsers do isomorphic-decode, and it's not defined by the
standard.  This CL removes this non-standard behavior.

This CL doesn't apply Blink's intent process. This is a very small
behavior change, and affects only 0.0016% of page views.

* Implementation
 - Add a flag to disable isomorphic-decode to
url::DecodeURLEscapeSequences(), and apply the new behavior only to
blink::LocalFrameView::ProcessUrlFragment().

 - Remove unnecessary UseCounters.


chromestatus: https://www.chromestatus.com/feature/4885685374812160
[1] https://infra.spec.whatwg.org/#isomorphic-decode

Change-Id: I17788e3697484701db66dbfef53bd1cae4598c51
Bug: 845824
Reviewed-on: https://chromium-review.googlesource.com/c/1333178
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: David Jacobo <djacobo@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#615419}