Fix the issue when Webdriver test suit retries even when the first run is green.
Currently all Webdriver test failures are expected failures until the two-way
sync with upstream github is created.

The bug is caused by the fact that expected != actual in the output Json
output while the is_unexpected (by default when this field is not set) field
is false. Comparing patch 9 where two fields are equal and the webdrvier test
suit didn't retry when green, to patch 10 where two fields are not equal and
webdriver test suits retried, I conclude that this is the cause of the bug
for 929689 and 936557

Bug: 929689, 936557
Change-Id: Iee0da2344ef1a6feff3ccde4746810f3c5ec8b76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1490897
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Kyle Ju <kyleju@chromium.org>
Cr-Commit-Position: refs/heads/master@{#637559}