Switch HttpServer::Delegate::OnWebSocketMessage to std::string.
Previously, this was a const std::string&, so in practice this
forces a copy. Now, the server std::moves the payload to the
delegate instance when calling OnWebSocketMessage.

I'm not updating client code more than necessary here, so there
are still copies e.g. in devtools_http_handler.cc. I plan to
resolve them a different way, by introducing a custom class
for devtools messages. But since net is a lower layer, it
seems better to do the string moves there and keep it minimal.

Change-Id: Ia74c7b67b19f2c6d80494889721b7201b2a3b5af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525405
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641310}