Fixed TakeElementScreenshot to crop large elements to the viewport
Fixed TakeElementScreenshot to truncate the dimensions of the screenshot
to be less than or equal to those of the viewport. Previously, elements
larger than the viewports would not be clipped during the screenshot.
Additionally added a test "testTakeElementScreenshotFrame" to
run_py_tests.py and an html file large_element.html that tests for
this bug. The html file itself is just a large red square greater than
the size of the viewport encompassed by a larger blue square and
white background. The test takes a screenshot of the html file and
ensures that the entirety of the screenshot is the red and no other
colors are present.

R=crouleau@google.com, zhanliang@google.com

Test: Run testTakeElementScreenshotFrame in run_py_tests.py
Bug: chromedriver:2940 b/134190465
Change-Id: I359268fdfd3e6a5cf5e6331cd5f6ec68eedd6c71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648735
Auto-Submit: Julian Kung <juliankung@google.com>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667674}