[ChromeDriver] Avoid cache interference
Use an object with a null prototype for the internal mapping of names to
cached values. This reduces the likelihood of unintended interaction
with the environment (e.g. when new enumerable prototypes are defined on
the Object prototype).

Update the testing code so that this change is verifiable.

Bug: 1014780
Change-Id: Ifbdc8d09c1049d0ec07f7a0099d774579423cdb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894707
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713496}