events: ozone: Remove implicit init from GetKeyboardLayoutEngine
GetKeyboardLayoutEngine() has the side effect of creating a stub keyboard
layout engine if there's no previous layout engine. Avoid this as it
shouldn't be necessary and it creates ordering dependencies between
tests. In particular, running EventConverterEvdevImplTest after
XkbLayoutEngineVkTest causes a crash when we enable use_xkbcommon in GN.
This shows up as a flaky failure on infra:

  flaky failures [ignored]:
  EventConverterEvdevImplTest.KeyPress

Removing the implicit initialization required fixes to a few components:

- KeyEvent tests must create a layout engine if they call GetDomKey or GetCharacter
- ozone/x11 needs to provide a layout engine to chromium. There was no
  layout engine support in that component at all. This needs followup
  work as Linux builds that need proper keyboard layout support need to
  use XkbKeyboardLayoutEngine rather than StubKeyboardLayoutEngine.
- Chromedriver needs to create a layout engine if run headlessly in X11
  (USE_X11) builds without an X display. This is a particularly unusual
  case as we've building ozone code in a non-ozone build.

This patch makes those components create a StubKeyboardLayoutEngine
explicitly, which is equivalent to the previous implicit behavior.

KeyEvent tests that rely on the global layout (because they call
GetDomKey or GetCharacter) now use ScopedKeyboardLayout, which has been
updated to create a StubKeyboardLayoutEngine under the hood. If more
control is required, a custom layout engine can be provided via
ScopedKeyboardLayoutEngine (ozone specific).

Bug: 1026406

Change-Id: I925ed2d0de880e19eaee41a2b36595fea467b6bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1928250
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Kevin Schoedel <kpschoedel@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718779}