[ChromeDriver] Dispatch touch events with their actual timestamp
When we are handling the consequence touch move events, sometimes, we
will coalesce them together, and we need the events' timestamp to
decide to which one dispatch first. I use the actual timestamp when
the event is generated and about to dispatch in
Input.DispatchTouchEvent function.

Bug: 1020674
Change-Id: I8a56f6f532df3ca6182c323c6f8c14f4825589e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1943107
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720508}