Reland "Support Wheel input source in Action API"
We will add a new input source type to simulate the mouse wheel scroll
to the Webdriver Action API, https://github.com/w3c/webdriver/pull/1522.

In this CL, we are adding the wheel input source in the ChromeDriver's
ExecutePerformActions function, which implements the Webdriver Action
API in Chrome and we will use Devtool's input protocol,
Input.dispatchMouseEvent with a type of "mouseWheel".

I added some waiting time to wait for the action to be executed to fix
the failure on Mac 14 test.

TBR=johnchen@chromium.org

Bug: 1040611
Change-Id: I348c71bacc32b403ba995c1c736b6bfb230b65aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337255
Reviewed-by: Lan Wei <lanwei@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794706}