Reland "Support triple click in Chromedriver PerformActions"
When we are using PerformActions from ChromeDriver to simulate three
mouse click, it only set the click count to 2, not 3, so this test fails
third_party/blink/web_tests/external/wpt/webdriver/tests/perform_actions
/pointer_tripleclick.py.

PerformActions should keep the mouse click count and decide if we should
increase the count or set to 1 when we have a mouse press based on the
interval of two mouse presses and difference between the current mouse
positions and the last one.

We should keep the click count in the session, and calculate the new
click count based on the count stored in the session.

Bug: 1119679
Change-Id: Iadac8e0f7dd13aebad6dc06c48c572426709a065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2368335
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Lan Wei <lanwei@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801082}