[Chromedriver] Add triple click tests in Chromedriver
When we are using Chromedriver ExecutePerformActions to simulate
three mouse click, it should set the mouse click count to 3, but if we
keep clicking, the click count should keep increasing on Windows and
Mac, reset to 1 otherwise. It should keep the mouse click count and
decide if we should increase the count or set to 1 when we have a mouse
press based on the interval of two mouse presses and difference between
the current mouse positions and the last one. I added three more tests
to verify these cases.

Bug: 1119679
Change-Id: I6f628e625e680870a77e18495eb0b4fab024fbbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2508987
Reviewed-by: Shengfa Lin <shengfa@google.com>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#823399}