Net: Save CNAME aliases to AddressList with aim of exposing to ad-tag.
The overall project aims to expose CNAME aliases to the
SubresourceFilter to aid in ad tagging and blocking.

This change stores the aliases (which are already
being extracted) in a new field of type
std::vector<std::string> in net::AddressList called `dns_aliases_`.

The first entry of this vector, if it exists, will be
the canonical name, and so this change also removes the previous
`canonical_name_` field and renames the canonical name getters and
setters with camel case, making them no longer inline, as now they have
compound logic.

This CL moreover provides the ability to merge lists of aliases to
the net::HostCache::Entry class, as well as some additional unit tests
and updates to existing ones.

Relevant tests:
net:net_unittests
--gtest_filter=*GetCanonicalNameWhenUnset*
--gtest_filter=*DnsAliases*
--gtest_filter=*CnameChain*TypeA*
--gtest_filter=*MergeEntries*

Bug: 1151047
Change-Id: Ie45cb4284df33053db36b364ea982cbfe3d6a618
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2561274
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#834439}