sheriff: Mass disable failing tests from linux-chromeos-rel
This is most of the tests that have failed on linux-chromeos-rel in
the last 200 builds (i.e. since the 4th of Feb). Due to the large
number of failures, most of which seem to be individually rare, I am
not attempting to determine which of these are fixed already on ToT.

Bug: 1177109, 1177111, 1177112, 1177113, 1177115
Bug: 1177120, 1177122, 1177126, 1177129, 1177130
Bug: 1177137, 1177138, 1177139, 1177140, 1177142
Bug: 1177117, 1177118, 1177131, 1177135, 1177146
Change-Id: I33c222287d9794cc5b166f400e846c9c23814705
AX-Relnotes: n/a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2690122
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Owners-Override: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#853028}