[CodeHealth] Remove deprecated SetDictionary calls in chromedriver
Replaces deprecated SetDictionary calls with the appropriate SetKey or
SetPath call. This also results in a change to take a moved Value instead
of a moved UniquePtr, thus most Set(Path|Key) calls were moved to the
old constructor spot, with the returned pointer used for the rest of the
method.

A few other deprecated calls were also updated as a result of this work.

Bug: 1187029
Change-Id: Idba3f975c7512630a8f33df30222fe6b24d58b0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3352401
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#953571}