Remove deprecated GetInteger methods in chromedriver
Removes uses of the deprecated DictionaryValue::GetInteger method in
chromedriver code. In *most* cases, this requires converting the int
that was previously used to an optional<int> and using that value
instead of the return value from the GetInteger method for any logic. A
few places, particularly in unit tests, make use of .value_or when there
is a clear "incorrect" value that can be assigned, or when the number is
already being validated.

A few places in window_commands that did not check the return value
either now add a check or force a ".value" call to trigger a CHECK, to
more closely match what seems to be the intended behavior.

This is a refactor with no intended behavior change (though a couple of
places ended up with simpler conditional checks).

Bug: 1187034
Change-Id: I555b6c57d07a16feaf6ce9c8eece2a8c6e792415
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3360229
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#954616}