ChromeDriver: Fix issues with shadow-dom fenced frames
The current implementation doesn't handle the case when a fenced frame
is appended. This is because events like childNodeInserted do not
include any shadowDom descendants of the fenced frame. This CL makes
DomTracker refetch the entire document tree (including shadow roots) and
rebuild the mappings when we encounter a fenced frame with incomplete
shadow dom.

Bug: chromedriver:4046
Change-Id: I4c8878a28023d423b23beef8327290945a3a9428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503635
Reviewed-by: Vladimir Nechaev <nechaev@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978720}