Eliminate base::LowerCaseEqualsASCII in /chrome/test/chromedriver
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=nechaev@chromium.org

Bug: 1326529
Change-Id: I5d2d613667302f9cc6008eacd4d24d361d56f174
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3655941
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1006337}