Renumber BiDi command ids
We want to distinguish responses to the user BiDi commands received
via different WebSocket channels and the infrastructural BiDi commands
issued by the ChromeDriver itself.
This is achieved via BiDi command re-numbering that takes into the
account the channel over which the command has arrived.
When the response is received from the BiDiMapper the initial command
id is restored in the response.

Bug: chromedriver:4181
Change-Id: I115d04a95361b529c56c85a33178dc52bc232d6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930595
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1056299}