Remove DevTools protocol "prompt" permission override option
The DevTools protocol allows permissions for websites to be
overridden[1]. As well as "granted"/"denied" the override could also
be for "prompt", which should have forced the permission re-requested
every time. In practice however, the "prompt" override wasn't working
correctly. The changes required to fix that proved to be too
invasive, so we decided[2] to remove the option instead.

1 - https://chromedevtools.github.io/devtools-protocol/tot/Browser/#method-setPermission
2 - https://crbug.com/1319943#c13

Change-Id: Iadb18066eb83d518d7ccc2ffbd89c8f6c24fac0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3683443
Reviewed-by: Vladimir Nechaev <nechaev@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Dave Vandyke <kzar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1067061}