Add DevToolsClient::RemoveListener() method.
Avoids some dangling ptr warnings while running chromedriver_unittests.

The JavaScriptDialogManager takes a client argument in its
constructor, and maintains an unowned reference to the client
for its lifetime. This means the client must outlive the manager.

The manager's constructor also adds the manager to the client, for
the client's lifetime. This creates a paradox where each must outlive
the other to avoid dangling a ptr.

Balance the addition in the constructor with a removal in the
destructor. Then move the declaration of locals in tests so that
when the test body finishes, the client outlives the manager as
locals are destructed in inverse order to their appearance.

Bug: 1401495
Change-Id: I129a3477436b0def1cd52f79676da4e579d10afe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4257886
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1116584}